use flash_attn_softmax in extend_attention
